Make sure SQL Injection can actually occur
The first version of this only passed in exactly 1 row was returned.

I intended this to be a test for 0 or 1, but SQL injection like
username=ZAP' OR 1=1; --
resulted in it returning all the rows for all the users.

I actually want that to be detected, so switch to returning just the
first row